requirements.txt
pandas
numpy
torch
tinkoff-invest